+++ module-common/src/main/java/com/yapp/sharefood/report/exception/NotDefineReportType.java
@@ -1,4 +1,4 @@
-package com.yapp.sharefood.report.domain;
+package com.yapp.sharefood.report.exception;
 
 public class NotDefineReportType extends RuntimeException {
     public static String NOT_DEFINE_REPORT_EXCEPTION_MSG = "정의되지 않은 신고 사유입니다.";
